feat: score and rank
5/27 개발

+++ Shared/Network/Player.cs
@@ -30,7 +30,7 @@ namespace Shared.Network
             return $"[Player] ClientId={ClientId}, Name={Name}";
         }
 
-        public int CurrentScore { get; set; }
+        public ushort CurrentScore { get; set; }
         //public int CurrentRank { get; set; } 
     }
 }
